Fixed a todo and did some reorganization
