+++ mala/descriptors/label_sublib/young.py
@@ -1,4 +1,8 @@
 from sympy.combinatorics import Permutation
+import math
+import numpy as np
+from mala.descriptors.tree_sorting import tree
+import itertools
 
 local_sigma_c_partitions = {
     1: [(1,)],
@@ -66,6 +70,39 @@ def is_row_sort(partitionfill):
         logi = tuple(sorted(orbit)) == orbit
         all_srt.append(logi)
     return all(all_srt)
+
+
+def group_vec_by_node(vec, nodes, remainder=None):
+    # vec_by_tups = [tuple([vec[node[0]],vec[node[1]]]) for node in nodes]
+    vec_by_tups = []
+    for node in nodes:
+        orbit_list = []
+        for inode in node:
+            orbit_list.append(vec[inode])
+        orbit_tup = tuple(orbit_list)
+        vec_by_tups.append(orbit_tup)
+    if remainder != None:
+        vec_by_tups = vec_by_tups + [tuple([vec[remainder]])]
+    return vec_by_tups
+
+
+def group_vec_by_orbits(vec, part):
+    ind_range = np.sum(part)
+    assert len(vec) == ind_range, "vector must be able to fit in the partion"
+    count = 0
+    by_orbits = []
+    for orbit in part:
+        orbit_vec = []
+        for i in range(orbit):
+            orbit_vec.append(vec[count])
+            count += 1
+        by_orbits.append(tuple(orbit_vec))
+    return tuple(by_orbits)
+
+
+def unique_perms(vec):
+    all_perms = [p for p in itertools.permutations(vec)]
+    return sorted(list(set(all_perms)))
 
 
 class YoungSubgroup:
